risolti alcuni problemi che causavano il troncamento dei valori decimali a numeri interi, laddove presenti, come nel caso dei valori di glicemia. cambiato il nome della chiave associata alla data di rilevamento della saturazione dell'ossigeno, perché sia conforme con la struttura gestita dal back-end.

+++ android/src/main/java/com/reactnative/samsunghealth/BloodGlucoseResultListener.java
@@ -113,7 +113,7 @@ public class BloodGlucoseResultListener implements
 
                         col = c.getColumnIndex(HealthConstants.BloodGlucose.GLUCOSE);
                         if (col > -1) {
-                            map.putDouble(HealthConstants.BloodGlucose.GLUCOSE, (double) c.getLong(col));
+                            map.putDouble(HealthConstants.BloodGlucose.GLUCOSE, (double) c.getFloat(col));
                         }
 
                         /*
